Refactor configuration so it is sane. Bunch of cleanup.
