+++ src/main/kotlin/FlacDatabase.kt
@@ -1,6 +1,5 @@
 package com.eigenholser.flac2mp3
 
-import org.jetbrains.exposed.dao.id.IntIdTable
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.transactions.TransactionManager
 import org.jetbrains.exposed.sql.transactions.transaction
