+++ src/main/kotlin/ImageScaler.kt
@@ -1,7 +1,6 @@
 package com.eigenholser.flac2mp3
 
 import ij.IJ
-import ij.ImagePlus
 import ij.process.ImageProcessor
 import kotlin.math.nextUp
 
@@ -10,25 +9,23 @@ enum class DestType {
 }
 
 object ImageScaler {
+    // TODO: Add to config?
     val coverFilename = "cover.jpg"
     val thumbFilename = "thumb.jpg"
     val destFormat = "jpg"
 
     private fun computeScaleFactor(xAxis: Int, srcSize: Int): Double {
-        println("${xAxis/srcSize.toDouble()}")
         return (xAxis/srcSize.toDouble())
     }
 
     private fun makeThumb(ip: ImageProcessor): ImageProcessor {
-        //TODO: Get resolution values from config file
-        val scaleFactor = computeScaleFactor(200, ip.width)
-        return ip.resize(200, (scaleFactor*ip.height).nextUp().toInt())
+        val scaleFactor = computeScaleFactor(Config.thumbnailResolution, ip.width)
+        return ip.resize(Config.thumbnailResolution, (scaleFactor*ip.height).nextUp().toInt())
     }
 
     private fun makeCover(ip: ImageProcessor): ImageProcessor {
-        //TODO: Get resolution values from config file
-        val scaleFactor = computeScaleFactor(1000, ip.width)
-        return ip.resize(1000, ((scaleFactor*ip.height).nextUp().toInt()))
+        val scaleFactor = computeScaleFactor(Config.coverResolution, ip.width)
+        return ip.resize(Config.coverResolution, ((scaleFactor*ip.height).nextUp().toInt()))
     }
 
     fun scaleImage(src: String, dest: String) {
@@ -43,7 +40,7 @@ object ImageScaler {
     }
 
     fun scaleImage(src: String, dest: String, destType: DestType) {
-        val imp = IJ.openImage("$src/album_art.png")
+        val imp = IJ.openImage("$src/${Config.albumArtFile}")
         val ip = imp.processor
 
         if (destType == DestType.THUMB) {
