+++ src/main/kotlin/main.kt
@@ -8,41 +8,25 @@ import java.nio.file.Files
 import java.nio.file.attribute.FileTime
 import kotlin.io.path.ExperimentalPathApi
 
+
+
 @ExperimentalPathApi
 fun main(args: Array<String>) {
-    val configFile = System.getenv("HOME") + "/flac2mp3.properties"
-    println("Reading from $configFile")
-
-    val config = ConfigFactory.parseFile(File(configFile))
-    val mp3Root = config.extract<String>("mp3_root")
-    println(mp3Root)
-    val flacRoot = config.extract<String>("flac_root")
-    println(flacRoot)
-    val albumArtFile = config.extract<String>("album_art.name")
-    println(albumArtFile)
-    val coverResolution = config.extract<Int>("album_art.resolution.cover")
-    println(coverResolution)
-    val thumbnailResolution = config.extract<Int>("album_art.resolution.thumb")
-    println(thumbnailResolution)
-
     val db = DbSettings.db
     FlacDatabase.createDatabase()
 
-    File(flacRoot)
+    File(Config.flacRoot)
         .walk()
         .filter {
             it.extension == "flac"
         }
         .forEach { file ->
-            println("flacfile: $file")
-            print("fsize: ${Files.getAttribute(file.toPath(), "size")}  ")
-            println("mtime: ${Files.getAttribute(file.toPath(), "lastModifiedTime")}")
             val flacfile = file.absolutePath
             val fsize = Files.getAttribute(file.toPath(), "size") as Long
             val mtime = Files.getAttribute(file.toPath(), "lastModifiedTime") as FileTime
             val tags = Tag.readFlacTags(flacfile)
             println(tags)
-            val flacRelativePath = flacfile.removePrefix("$flacRoot/")
+            val flacRelativePath = flacfile.removePrefix("${Config.flacRoot}/")
 
             //val flacRow = FlacDatabase.getByCddbAndTrack(tags.cddb, tags.track)
 
@@ -57,15 +41,15 @@ fun main(args: Array<String>) {
     var nextAlbum = ""
 
     FlacDatabase.getAllFlacRows().forEach {
-        val flacFileAbsolute = File("$flacRoot/${it[Flac.flacfile]}")
+        val flacFileAbsolute = File("${Config.flacRoot}/${it[Flac.flacfile]}")
         val flacAlbumPathAbsolute = File(flacFileAbsolute.toString().removeSuffix("/${flacFileAbsolute.name}"))
-        val flacAlbumArtFile = File("$flacAlbumPathAbsolute/$albumArtFile")
+        val flacAlbumArtFile = File("$flacAlbumPathAbsolute/${Config.albumArtFile}")
         val flacFileTrackName = flacFileAbsolute.name
         val currentAlbum = flacAlbumPathAbsolute.toString()
-            .removePrefix("$flacRoot/")
+            .removePrefix("${Config.flacRoot}/")
             .removeSuffix("/${flacFileTrackName}")
 
-        val mp3AlbumPathAbsolute = File("$mp3Root/$currentAlbum").toPath()
+        val mp3AlbumPathAbsolute = File("${Config.mp3Root}/$currentAlbum").toPath()
         val mp3FileAbsolute = File("$mp3AlbumPathAbsolute/${flacFileAbsolute.nameWithoutExtension}.mp3")
 
         if (!switchAlbum && currentAlbum != nextAlbum) {
